+++ robot_simulator/grid/board.py
@@ -4,7 +4,7 @@
 from robot_simulator.grid.positioning import Point, Rect
 
 
-class Board(object):
+class Board:
     def __init__(self, size_x, size_y):
         bottom_left = Point(0, 0)
         top_right = Point(size_x - 1, size_y - 1)